update types - compile warning for connected multi
